updating godocs and code organization
