+++ modules/apps/27-interchain-accounts/types/account_test.go
@@ -49,80 +49,6 @@ func (suite *TypesTestSuite) TestGenerateAddress() {
 	suite.Require().NotEmpty(accAddr)
 }
 
-func (suite *TypesTestSuite) TestGeneratePortID() {
-	var (
-		path  *ibctesting.Path
-		owner = TestOwnerAddress
-	)
-
-	testCases := []struct {
-		name     string
-		malleate func()
-		expValue string
-		expPass  bool
-	}{
-		{
-			"success",
-			func() {},
-			fmt.Sprint(types.VersionPrefix, types.Delimiter, "0", types.Delimiter, "0", types.Delimiter, TestOwnerAddress),
-			true,
-		},
-		{
-			"success with non matching connection sequences",
-			func() {
-				path.EndpointA.ConnectionID = "connection-1"
-			},
-			fmt.Sprint(types.VersionPrefix, types.Delimiter, "1", types.Delimiter, "0", types.Delimiter, TestOwnerAddress),
-			true,
-		},
-		{
-			"invalid connectionID",
-			func() {
-				path.EndpointA.ConnectionID = "connection"
-			},
-			"",
-			false,
-		},
-		{
-			"invalid counterparty connectionID",
-			func() {
-				path.EndpointB.ConnectionID = "connection"
-			},
-			"",
-			false,
-		},
-		{
-			"invalid owner address",
-			func() {
-				owner = "    "
-			},
-			"",
-			false,
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		suite.Run(tc.name, func() {
-			suite.SetupTest() // reset
-			path = ibctesting.NewPath(suite.chainA, suite.chainB)
-			suite.coordinator.Setup(path)
-
-			tc.malleate()
-
-			portID, err := types.GeneratePortID(owner, path.EndpointA.ConnectionID, path.EndpointB.ConnectionID)
-
-			if tc.expPass {
-				suite.Require().NoError(err, tc.name)
-				suite.Require().Equal(tc.expValue, portID)
-			} else {
-				suite.Require().Error(err, tc.name)
-				suite.Require().Empty(portID)
-			}
-		})
-	}
-}
-
 func (suite *TypesTestSuite) TestInterchainAccount() {
 	pubkey := secp256k1.GenPrivKey().PubKey()
 	addr := sdk.AccAddress(pubkey.Address())
