+++ modules/apps/27-interchain-accounts/types/keys.go
@@ -2,12 +2,6 @@ package types
 
 import (
 	"fmt"
-	"strconv"
-	"strings"
-
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
-	porttypes "github.com/cosmos/ibc-go/v2/modules/core/05-port/types"
 )
 
 const (
@@ -31,10 +25,6 @@ const (
 
 	// Delimiter is the delimiter used for the interchain accounts version string
 	Delimiter = "."
-
-	// ControllerPortFormat is the expected port identifier format to which controller chains must conform
-	// See (TODO: Link to spec when updated)
-	ControllerPortFormat = "<app-version>.<controller-conn-seq>.<host-conn-seq>.<owner>"
 )
 
 var (
@@ -61,36 +51,4 @@ func KeyOwnerAccount(portID string) []byte {
 // KeyPort creates and returns a new key used for port store operations
 func KeyPort(portID string) []byte {
 	return []byte(fmt.Sprintf("%s/%s", PortKeyPrefix, portID))
-}
-
-// ParseControllerConnSequence attempts to parse the controller connection sequence from the provided port identifier
-// The port identifier must match the controller chain format outlined in (TODO: link spec), otherwise an empty string is returned
-func ParseControllerConnSequence(portID string) (uint64, error) {
-	s := strings.Split(portID, Delimiter)
-	if len(s) != 4 {
-		return 0, sdkerrors.Wrap(porttypes.ErrInvalidPort, "failed to parse port identifier")
-	}
-
-	seq, err := strconv.ParseUint(s[1], 10, 64)
-	if err != nil {
-		return 0, sdkerrors.Wrapf(err, "failed to parse connection sequence (%s)", s[1])
-	}
-
-	return seq, nil
-}
-
-// ParseHostConnSequence attempts to parse the host connection sequence from the provided port identifier
-// The port identifier must match the controller chain format outlined in (TODO: link spec), otherwise an empty string is returned
-func ParseHostConnSequence(portID string) (uint64, error) {
-	s := strings.Split(portID, Delimiter)
-	if len(s) != 4 {
-		return 0, sdkerrors.Wrap(porttypes.ErrInvalidPort, "failed to parse port identifier")
-	}
-
-	seq, err := strconv.ParseUint(s[2], 10, 64)
-	if err != nil {
-		return 0, sdkerrors.Wrapf(err, "failed to parse connection sequence (%s)", s[2])
-	}
-
-	return seq, nil
 }
